pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-classifier"
version = "0.1.0"
description = "A PoC for image content classification using multimodal LLMs"
authors = [
    {name = "User", email = "user@example.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.4.0",
    "pyyaml>=6.0.1",
    "aiohttp>=3.9.0",
    "pillow>=10.0.0",
    "jinja2>=3.1.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
openai = [
    "openai>=1.3.0",
]
anthropic = [
    "anthropic>=0.7.0",
]
google = [
    "google-generativeai>=0.3.0",
]
all = [
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "google-generativeai>=0.3.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]